[doyto-query-reactive] upgrade versions

+++ doyto-query-r2dbc/src/main/java/win/doyto/query/r2dbc/R2dbcOperations.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright © 2019-2022 Forb Yuan
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package win.doyto.query.r2dbc;
 
 import reactor.core.publisher.Flux;
@@ -17,9 +33,9 @@ public interface R2dbcOperations {
 
     <I> Mono<I> insert(SqlAndArgs sqlAndArgs, String idColumn, Class<I> idClass);
 
-    Mono<Integer> update(SqlAndArgs sqlAndArgs);
+    Mono<Long> update(SqlAndArgs sqlAndArgs);
 
-    default Mono<Integer> update(String sql, Object... args) {
+    default Mono<Long> update(String sql, Object... args) {
         return update(new SqlAndArgs(sql, args));
     }
